feat: LineSectionAcceptanceTest(기존) 제거 / ModifiedLineSectionAcceptanceTest -> LineSectionAcceptanceTest 이름변경
